Add NextTopLoader to layout
Integrated the 'nextjs-toploader' package into the project. This was done by adding it as a dependency in the package.json file and importing it into the layout.tsx file. The TopLoader component is now rendered in our RootLayout function with custom properties such as color, height, initial position, easing, and speed.