Update branding in SidebarHeader and Footer
The branding text in the SidebarHeader and Footer components has been updated. The previous brand name 'ESTIMATIC' has been replaced with 'CostCraft', and the author's name 'Kamil Woźniacki' is now represented as 'Constructive Creativity'. These changes are reflected both in the header of the sidebar and in the copyright notice of the footer.