Add Google icon to login button
A new SVG file for the Google icon has been added. This icon is now displayed on the 'Continue with Google' button in the LoginCard component. The button text has also been updated from 'Login with Google' to 'Continue with Google'. Additionally, an email icon was added to the 'Send Magic Link' button.