Refactor Google Drive folder creation function
The Google Drive folder creation function has been updated to accept an additional parameter for client email. This is used to set permissions on the created folder, allowing the specified user read access. The EstimateRequestForm component has also been adjusted to pass in the user's email when creating a new folder.