Add OTP sent confirmation and loading state
Introduced a new page to confirm that the One-Time Password (OTP) has been sent. Also added a loading state in the login form when sending the magic link. The site URL is now dynamically set based on environment variables or defaults to localhost for development. After successfully sending the OTP, users are redirected to the newly created confirmation page.