Add QuoteStepInfoAccordion to new quote page
A new component, QuoteStepInfoAccordion, has been added to the new quote page. This accordion-style component displays additional information about each step in the quote creation process. The information is dynamically loaded based on the current step of the process. A new file for this component has also been created and imported into the main page file.